// frontend/src/api/generated/gateways/gateways.ts
export type gatewaysStatusApiV1GatewaysStatusGetResponse200 = {
  data: GatewaysStatusResponse
  status: 200
}

export type gatewaysStatusApiV1GatewaysStatusGetResponse422 = {
  data: HTTPValidationError
  status: 422
}
    
export type gatewaysStatusApiV1GatewaysStatusGetResponseSuccess = (gatewaysStatusApiV1GatewaysStatusGetResponse200) & {
  headers: Headers;
};
export type gatewaysStatusApiV1GatewaysStatusGetResponseError = (gatewaysStatusApiV1GatewaysStatusGetResponse422) & {
  headers: Headers;
};

export type gatewaysStatusApiV1GatewaysStatusGetResponse = (gatewaysStatusApiV1GatewaysStatusGetResponseSuccess | gatewaysStatusApiV1GatewaysStatusGetResponseError)
// ...
export type getGatewayApiV1GatewaysGatewayIdGetResponse200 = {
  data: GatewayRead
  status: 200
}

export type getGatewayApiV1GatewaysGatewayIdGetResponse422 = {
  data: HTTPValidationError
  status: 422
}
    
export type getGatewayApiV1GatewaysGatewayIdGetResponseSuccess = (getGatewayApiV1GatewaysGatewayIdGetResponse200) & {
  headers: Headers;
};
export type getGatewayApiV1GatewaysGatewayIdGetResponseError = (getGatewayApiV1GatewaysGatewayIdGetResponse422) & {
  headers: Headers;
};

export type getGatewayApiV1GatewaysGatewayIdGetResponse = (getGatewayApiV1GatewaysGatewayIdGetResponseSuccess | getGatewayApiV1GatewaysGatewayIdGetResponseError)