// frontend/src/api/generated/agents/agents.ts
export type listAgentsApiV1AgentsGetResponse200 = {
  data: LimitOffsetPageTypeVarCustomizedAgentRead
  status: 200
}

export type listAgentsApiV1AgentsGetResponse422 = {
  data: HTTPValidationError
  status: 422
}
    
export type listAgentsApiV1AgentsGetResponseSuccess = (listAgentsApiV1AgentsGetResponse200) & {
  headers: Headers;
};
export type listAgentsApiV1AgentsGetResponseError = (listAgentsApiV1AgentsGetResponse422) & {
  headers: Headers;
};

export type listAgentsApiV1AgentsGetResponse = (listAgentsApiV1AgentsGetResponseSuccess | listAgentsApiV1AgentsGetResponseError)